The place where I found this code - http://www.mrtc.mdh.se/projects/wcet/wcet_bench/edn/edn.c had made some changes which effect the output.
Because cast has higher precedence than *, the (short) casts are
nops, applying as they do to the left argument of a multiply, rather
than the result of the multiply (which is probably what is intended).

This reverts that change, to match the code at
http://www.edn.com/archives/1997/060597/12df_02.htm - to get something
that we can compare against.



git-svn-id: svn://localhost/svn/uclinux-dist/trunk@7796 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ user/blkfin-apps/edn/edn.c
@@ -160,7 +160,7 @@ void
 jpegdct(short *d, const short *r)
 {
 	int t[12];
-	int i, j, k, m, n, p;
+	short i, j, k, m, n, p;
 
 	for (k = 1, m = 0, n = 13, p = 8;
 		k <= 8;
@@ -177,8 +177,8 @@ jpegdct(short *d, const short *r)
 			d[0] = (t[8] + t[10]) >> m;
 			d[4 * k] = (t[8] - t[10]) >> m;
 			t[8] = (short) (t[11] + t[9]) * r[10];
-			d[2 * k] = t[8] + (short) ((t[9] * r[9]) >> n);
-			d[6 * k] = t[8] + (short) ((t[11] * r[11]) >> n);
+			d[2 * k] = t[8] + (short) t[9] * r[9] >> n;
+			d[6 * k] = t[8] + (short) t[11] * r[11] >> n;
 			t[0] = (short) (t[4] + t[7]) * r[2];
 			t[1] = (short) (t[5] + t[6]) * r[0];
 			t[2] = t[4] + t[6];
@@ -186,10 +186,10 @@ jpegdct(short *d, const short *r)
 			t[8] = (short) (t[2] + t[3]) * r[8];
 			t[2] = (short) t[2] * r[1] + t[8];
 			t[3] = (short) t[3] * r[3] + t[8];
-			d[7 * k] = (short) (t[4] * r[4] + t[0] + t[2]) >> n;
-			d[5 * k] = (short) (t[5] * r[6] + t[1] + t[3]) >> n;
-			d[3 * k] = (short) (t[6] * r[5] + t[1] + t[2]) >> n;
-			d[1 * k] = (short) (t[7] * r[7] + t[0] + t[3]) >> n;
+			d[7 * k] = (short) t[4] * r[4] + t[0] + t[2] >> n;
+			d[5 * k] = (short) t[5] * r[6] + t[1] + t[3] >> n;
+			d[3 * k] = (short) t[6] * r[5] + t[1] + t[2] >> n;
+			d[1 * k] = (short) t[7] * r[7] + t[0] + t[3] >> n;
 		}
 	}
 }
